Uploaded the first test to the Update decoder - will probably not work as I have not tested it yet! Theoretically however it has a chance of working already! The encoder call is not updated and will probably break the decoder - we have to make sure that they will work together!
New Update structure would be:
(CellPosition) SPACER ObjectOnCell:ObjectID:Inventory

The Inventory will only be used if said ObjectOnCell is a Robot. If that is the case, an example would be (I will use "_" for the DEFAULT SPACER to make it more readable)":
(10,2)_Robot:1:Ore:2
(10,2) is the cell position.
Robot is the Robot in question.
"1" is the teamID of the robot.
"Ore" is the inventory
"2" is the type of ore - its ID.